refactor: move cluster related actions from connectionStroe to clusterStore
