+++ src/views/manage/components/shared-manage.vue
@@ -11,12 +11,12 @@
 
     <div v-if="indexShards" class="shard-statistic-container">
       <h3 class="shard-statistic-title">
-        <span>{{ indexShards.index }}</span
-        ><span
-          >shards: {{ indexShards.shards.filter(shard => shard.prirep === 'p').length }}/{{
+        <span>{{ indexShards.index }}</span>
+        <span>
+          shards: {{ indexShards.shards.filter(shard => shard.prirep === 'p').length }}/{{
             indexShards.shards.filter(shard => shard.prirep === 'r').length
-          }}, unassigned: {{ indexShards.shards.filter(shard => !shard.node).length }}</span
-        >
+          }}, unassigned: {{ indexShards.shards.filter(shard => !shard.node).length }}
+        </span>
       </h3>
       <div class="shard-list-scrollbar-box">
         <n-scrollbar style="height: 100%">
@@ -44,7 +44,7 @@
                 <n-tag :type="shardsDetail.tagType">
                   {{ shardsDetail.content }}
                   <template #icon>
-                    <n-icon :component="shardsDetail.icon" />
+                    <n-icon :component="shardsDetail.icon()" />
                   </template>
                 </n-tag>
               </template>
@@ -59,29 +59,29 @@
 
 <script setup lang="ts">
 import { storeToRefs } from 'pinia';
-import { Shard, ShardState, useConnectionStore } from '../../../store';
+import { Shard, ShardState, useClusterManageStore } from '../../../store';
 import prettyBytes from 'pretty-bytes';
 import {
-  Document,
-  VmdkDisk,
-  Insert,
-  QueryQueue,
-  Layers,
-  ShapeExcept,
-  SearchLocate,
-  Application,
   AiResults,
-  Version,
-  WarningAlt,
-  Rotate360,
+  Application,
+  Document,
+  Insert,
   LaunchStudy1,
+  Layers,
+  QueryQueue,
+  Rotate360,
+  SearchLocate,
+  ShapeExcept,
+  Version,
+  VmdkDisk,
+  WarningAlt,
 } from '@vicons/carbon';
 import { Memory } from '@vicons/fa';
 import { NButton } from 'naive-ui';
 
-const connectionStore = useConnectionStore();
-const { fetchNodes, fetchShards, getShardState } = connectionStore;
-const { established } = storeToRefs(connectionStore);
+const clusterManageStore = useClusterManageStore();
+const { fetchNodes, fetchShards, getShardState } = clusterManageStore;
+const { nodes, shards } = storeToRefs(clusterManageStore);
 
 type NodeWithShard = {
   columns: Array<{ title: string; key: string }>;
@@ -104,11 +104,10 @@ const indexShards = ref<{
 }>();
 
 const refreshShards = async (): Promise<NodeWithShard> => {
-  const nodes = established.value?.rawClusterState?.nodes.instances;
-  if (!nodes) return { columns: [], data: [] };
+  if (!nodes.value) return { columns: [], data: [] };
 
-  const shards = await fetchShards();
-  const columns = [{ name: 'index' }, { name: 'unassigned' }, ...nodes].map(column => ({
+  await fetchShards();
+  const columns = [{ name: 'index' }, { name: 'unassigned' }, ...nodes.value].map(column => ({
     title: column.name,
     key: column.name,
     render(row: { [key: string]: Array<Shard> }) {
@@ -124,19 +123,21 @@ const refreshShards = async (): Promise<NodeWithShard> => {
             onClick: () => handleShardClick(row as unknown as Shard),
             class: 'shard-box',
           },
-          `${shard.prirep}${shard.shard}`,
+          {
+            default: () => `${shard.prirep}${shard.shard}`,
+          },
         ),
       );
     },
   }));
 
-  const data = Array.from(new Set(shards?.map(shard => shard.index))).map(index => ({
+  const data = Array.from(new Set(shards.value?.map(shard => shard.index))).map(index => ({
     index,
-    unassigned: shards?.filter(shard => !shard.node && shard.index === index),
-    ...(nodes || []).reduce(
+    unassigned: shards.value?.filter(shard => !shard.node && shard.index === index),
+    ...(nodes.value || []).reduce(
       (acc, node) => ({
         ...acc,
-        [node.name]: shards?.filter(shard => shard.node && shard.index === index),
+        [node.name]: shards.value?.filter(shard => shard.node && shard.index === index),
       }),
       {},
     ),
@@ -151,13 +152,13 @@ const handleShardClick = async (shard: Shard) => {
     ...shard,
     details: [
       {
-        icon: () => h(Document),
+        icon: () => Document,
         content: `docs: ${shard.docs.count}`,
         desc: 'docs',
         tagType: 'success',
       },
       {
-        icon: () => h(VmdkDisk),
+        icon: () => VmdkDisk,
         content: `size: ${prettyBytes(shard.store.size)}, dataset: ${prettyBytes(
           shard.dataset.size,
         )}`,
@@ -165,49 +166,49 @@ const handleShardClick = async (shard: Shard) => {
         tagType: 'success',
       },
       {
-        icon: () => h(Memory),
+        icon: () => Memory,
         content: `size: ${prettyBytes(shard.completion.size)}`,
         desc: 'completion',
         tagType: 'success',
       },
       {
-        icon: () => h(Memory),
+        icon: () => Memory,
         content: `memory_size: ${prettyBytes(shard.fielddata.memorySize)}, evictions: ${shard.fielddata.evictions}`,
         desc: 'fielddata',
         tagType: 'success',
       },
       {
-        icon: () => h(Layers),
+        icon: () => Layers,
         content: `memory_size: ${prettyBytes(shard.queryCache.memorySize)}, evictions: ${shard.queryCache.evictions}`,
         desc: 'query_cache',
         tagType: 'success',
       },
       {
-        icon: () => h(QueryQueue),
+        icon: () => QueryQueue,
         content: `success: ${shard.get.existsTotal}, ${shard.get.existsTime} failure: ${shard.get.missingTotal}, ${shard.get.missingTime}`,
         desc: 'GET OPERATION',
         tagType: 'success',
       },
       {
-        icon: () => h(Insert),
+        icon: () => Insert,
         content: `index: ${shard.indexing.indexTime} delete: ${shard.indexing.deleteTotal}, ${shard.indexing.deleteTime} failures: ${shard.indexing.indexFailed}`,
         desc: 'INDEXING OPERATION',
         tagType: 'success',
       },
       {
-        icon: () => h(SearchLocate),
+        icon: () => SearchLocate,
         content: `fetch: ${shard.search.fetchTotal}/${shard.search.fetchTime}, query: ${shard.search.queryTotal}/${shard.search.queryTime}, scroll: ${shard.search.scrollTotal}/${shard.search.scrollTime}, open: ${shard.search.openContexts}`,
         desc: 'SEARCH OPERATION',
         tagType: 'success',
       },
       {
-        icon: () => h(ShapeExcept),
+        icon: () => ShapeExcept,
         content: `total: ${shard.merges.total}, size: ${prettyBytes(shard.merges.totalSize)}, docs: ${shard.merges.totalDocs} time: ${shard.merges.totalTime}`,
         desc: 'MERGES OPERATION',
         tagType: 'success',
       },
       {
-        icon: () => h(Application),
+        icon: () => Application,
         content: `count: ${shard.segments.count}/${prettyBytes(shard.segments.memory)}, writer: ${prettyBytes(
           shard.segments.indexWriterMemory,
         )}, version_map: ${prettyBytes(shard.segments.versionMapMemory)}, fixed_bitset: ${prettyBytes(
@@ -218,32 +219,32 @@ const handleShardClick = async (shard: Shard) => {
       },
 
       {
-        icon: () => h(Rotate360),
+        icon: () => Rotate360,
         content: `total: ${shard.refresh.total}, time: ${shard.refresh.time}`,
         desc: 'refresh',
         tagType: 'success',
       },
       {
-        icon: () => h(LaunchStudy1),
+        icon: () => LaunchStudy1,
         content: `total: ${shard.flush.total}, time: ${shard.flush.totalTime}`,
         desc: 'flush',
         tagType: 'success',
       },
       {
-        icon: () => h(Version),
+        icon: () => Version,
         content: `max: ${shard.seqNo.max}, global: ${shard.seqNo.globalCheckpoint}, local: ${shard.seqNo.localCheckpoint}`,
         desc: 'seq_no',
         tagType: 'success',
       },
       {
-        icon: () => h(AiResults),
+        icon: () => AiResults,
         content: `total: ${shard.suggest.total},time: ${shard.suggest.time}`,
         desc: 'suggest',
         tagType: 'success',
       },
       shard.unassigned.at
         ? {
-            icon: () => h(WarningAlt),
+            icon: () => WarningAlt,
             content: `details: ${shard.unassigned.details}, reason: ${shard.unassigned.reason}, for: ${shard.unassigned.for} at: ${shard.unassigned.at}`,
             desc: 'unassigned',
             tagType: 'warning',
