+++ src/views/manage/index.vue
@@ -3,7 +3,7 @@
     <tool-bar @switch-manage-tab="handleManageTabChange" />
     <cluster-state
       class="state-container"
-      :cluster="established?.rawClusterState || null"
+      :cluster="cluster"
       v-if="activeTab === $t('manage.cluster')"
     />
     <node-state class="state-container" v-if="activeTab === $t('manage.nodes')" />
@@ -15,22 +15,23 @@
 <script setup lang="ts">
 import ToolBar from './components/tool-bar.vue';
 import ClusterState from './components/cluster-state.vue';
-import { useConnectionStore } from '../../store';
+import { useClusterManageStore } from '../../store';
 import { storeToRefs } from 'pinia';
 import NodeState from './components/node-state.vue';
 import SharedManage from './components/shared-manage.vue';
 import { lang } from '../../lang';
 import IndexManage from './components/index-manage.vue';
+
 const activeTab = ref(lang.global.t('manage.cluster'));
 
-const connectionStore = useConnectionStore();
-const { fetchClusterState } = connectionStore;
+const clusterManageStore = useClusterManageStore();
+const { fetchCluster } = clusterManageStore;
+const { cluster } = storeToRefs(clusterManageStore);
 
-const { established } = storeToRefs(connectionStore);
 const handleManageTabChange = (tab: string) => {
   activeTab.value = tab;
 };
-fetchClusterState();
+fetchCluster().catch(() => {});
 </script>
 
 <style lang="scss" scoped>
@@ -39,6 +40,7 @@ fetchClusterState();
   height: 100%;
   display: flex;
   flex-direction: column;
+
   .state-container {
     flex: 1;
     height: 0;
